[SofaSceneAssist] Move the whole directory into the PSL one to make things clear.
