+++ applications/plugins/CMakeLists.txt
@@ -38,7 +38,7 @@ sofa_add_plugin(RigidScale RigidScale)
 sofa_add_plugin(LeapMotion LeapMotion)
 sofa_add_plugin(Geomagic Geomagic)
 sofa_add_plugin(CImgPlugin CImgPlugin ON) # ON by default
-sofa_add_plugin(SofaSceneAssist SofaSceneAssist ON) # ON by default
+sofa_add_plugin(PSL PSL ON) # ON by default
 
 #if((${CMAKE_CXX_COMPILER_ID} STREQUAL "GNU") AND (${CMAKE_SYSTEM_NAME} MATCHES "Linux"))
 #    sofa_add_plugin(SofaPardisoSolver SofaPardisoSolver) # SofaPardisoSolver is only available under linux with gcc
