+++ applications/plugins/PSL/examples/HiTec3.mtl
