+++ applications/plugins/PSL/examples/HiTec5b.mtl
