+++ applications/plugins/PSL/examples/struct.obj
