+++ applications/plugins/PSL/examples/test_rigidification.psl.txt
