+++ applications/plugins/PSL/examples/test_rigidification.pyson
